requirements.txt
python-dotenv==1.0.0
transformers==4.35.2
diffusers==0.24.0
datasets==2.15.0
langchain==0.0.348
streamlit==1.29.0
config==0.5.1
soundfile==0.12.1
torch==2.1.1
accelerate==0.25.0 
openai==1.3.8
sentencepiece==0.1.99 


